pyproject.toml
[tool.poetry]
name = "random-question"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
uvicorn = {extras = ["standart"], version = "^0.23.2"}
pydantic-settings = "^2.0.3"
sqlalchemy = "^2.0.22"
alembic = "^1.12.0"
asyncpg = "^0.28.0"
aiohttp = "^3.8.6"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
isort = "^5.12.0"
black = "^23.10.0"
pre-commit = "^3.5.0"

[tool.black]
line_length = 79

[tool.isort]
profile = "black"
line_length = 79
skip_glob = ["alembic/*.py", "**/versions/*.py"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
